refactor: rename folder and move files
